+++ src/solver/puzzle.rs
@@ -2,6 +2,6 @@ use crate::grid::Grid;
 use crate::constraints::constraint_set::ConstraintSet;
 
 pub struct Puzzle {
-    pub grid: Grid<Option<usize>>,
+    pub grid: Grid<Option<u8>>,
     pub constraint_set: ConstraintSet,
 }
